2-2-1

+++ GameEngine.h
@@ -14,7 +14,7 @@ private:
 	unique_ptr<ScreenManager> m_ScreenManager;
 
 	float m_FPS = 0;
-	Vector2f m_Resolution;
+	Vector2i m_Resolution;
 
 	void handleInput();
 	void update();
